Move sql functions from Deed to Deed Service
